+++ src/pathGenerator/Gui.java
@@ -485,7 +485,7 @@ public class Gui {
 		lblMotionProfileGenerator.setBounds(109, 29, 365, 64);
 		panel.add(lblMotionProfileGenerator);
 		
-		JLabel lblVersion = new JLabel("Version 2.2.0");
+		JLabel lblVersion = new JLabel("Version 2.3.0");
 		lblVersion.setFont(new Font("Arial", Font.PLAIN, 14));
 		lblVersion.setBounds(82, 104, 85, 14);
 		panel.add(lblVersion);