+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [2.3.0] - 2018-2-22
 ### Added
-- Added ability to choose between different units
+- Added ability to choose between different units (Feet and Meters)
 
 ### Fixed
 - Bug where a user could get stuck in point update mode